requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mlxtend
networkx
python-docx
Unidecode